Create README.md
Write a function to find the factorial using recursion.
Write a function to find the factorial using tail recursion.
Create a method to log. Take 2 arguments message as string and level as a string and then print something like: "[level]: message"
In the above method, if the level is not provided use "INFO" by default. (Hint: Default Argument)